rm effect/workflows folder
